# Generate new Equals overload to avoid boxing for structural comparison