--- a/src/Compiler/Checking/AugmentWithHashCompare.fsi
+++ b/src/Compiler/Checking/AugmentWithHashCompare.fsi
@@ -7,6 +7,13 @@ open FSharp.Compiler
 open FSharp.Compiler.TypedTree
 open FSharp.Compiler.TcGlobals
 
+[<NoEquality; NoComparison; StructuredFormatDisplay("{DebugText}")>]
+type EqualityWithComparerAugmentation =
+    { GetHashCode: Val
+      GetHashCodeWithComparer: Val
+      EqualsWithComparer: Val
+      EqualsExactWithComparer: Val }
+
 val CheckAugmentationAttribs: bool -> TcGlobals -> Import.ImportMap -> Tycon -> unit
 
 val TyconIsCandidateForAugmentationWithCompare: TcGlobals -> Tycon -> bool
@@ -21,7 +28,7 @@ val MakeValsForCompareWithComparerAugmentation: TcGlobals -> TyconRef -> Val
 
 val MakeValsForEqualsAugmentation: TcGlobals -> TyconRef -> Val * Val
 
-val MakeValsForEqualityWithComparerAugmentation: TcGlobals -> TyconRef -> Val * Val * Val
+val MakeValsForEqualityWithComparerAugmentation: TcGlobals -> TyconRef -> EqualityWithComparerAugmentation
 
 val MakeBindingsForCompareAugmentation: TcGlobals -> Tycon -> Binding list