--- a/src/Compiler/TypedTree/TypedTree.fs
+++ b/src/Compiler/TypedTree/TypedTree.fs
@@ -1246,7 +1246,7 @@ type Entity =
     member x.GeneratedHashAndEqualsValues = x.TypeContents.tcaug_equals
 
     /// Gets all implicit hash/equals/compare methods added to an F# record, union or struct type definition.
-    member x.AllGeneratedValues = 
+    member x.AllGeneratedInterfaceImplsAndOverrides = 
         [ match x.GeneratedCompareToValues with 
           | None -> ()
           | Some (vref1, vref2) -> yield vref1; yield vref2
@@ -1258,7 +1258,8 @@ type Entity =
           | Some (vref1, vref2) -> yield vref1; yield vref2
           match x.GeneratedHashAndEqualsWithComparerValues with
           | None -> ()
-          | Some (vref1, vref2, vref3) -> yield vref1; yield vref2; yield vref3 ]
+          // vref4 is compiled as a sealed instance member, not an interface impl or an override.
+          | Some (vref1, vref2, vref3, _) -> yield vref1; yield vref2; yield vref3 ]
     
 
     /// Gets the data indicating the compiled representation of a type or module in terms of Abstract IL data structures.
@@ -1398,10 +1399,10 @@ type TyconAugmentation =
       /// of Object.Equals or if the type doesn't override Object.Equals implicitly. 
       mutable tcaug_equals: (ValRef * ValRef) option
 
-      /// This is the value implementing the auto-generated comparison
+      /// This is the value implementing the auto-generated equality
       /// semantics if any. It is not present if the type defines its own implementation
-      /// of IStructuralEquatable or if the type doesn't implement IComparable implicitly.
-      mutable tcaug_hash_and_equals_withc: (ValRef * ValRef * ValRef) option                                    
+      /// of IStructuralEquatable or if the type doesn't override Object.Equals implicitly.
+      mutable tcaug_hash_and_equals_withc: (ValRef * ValRef * ValRef * ValRef option) option                                    
 
       /// True if the type defined an Object.GetHashCode method. In this 
       /// case we give a warning if we auto-generate a hash method since the semantics may not match up