--- a/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals11.fsx
+++ b/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals11.fsx
@@ -1,0 +1,6 @@
+module EqualsMicroPerfAndCodeGenerationTests = 
+
+    [<Struct>]
+    type SomeUnion = SomeUnion of int * int
+
+    SomeUnion (1, 2) = SomeUnion (2, 3) |> ignore