--- a/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals12.fsx
+++ b/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals12.fsx
@@ -1,0 +1,6 @@
+module EqualsMicroPerfAndCodeGenerationTests = 
+
+    [<Struct>]
+    type SomeRecord = { V: int; U: int }
+
+    { V = 1; U = 2 } = { V = 2; U = 3 } |> ignore