--- a/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals13.fsx
+++ b/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals13.fsx
@@ -1,0 +1,8 @@
+module EqualsMicroPerfAndCodeGenerationTests = 
+
+    [<Struct>]
+    type SomeGenericStruct<'T>(v: 'T, u: int) =
+        member _.V = v
+        member _.U = u
+
+    SomeGenericStruct(1, 2) = SomeGenericStruct(2, 3) |> ignore