--- a/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals14.fsx
+++ b/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals14.fsx
@@ -1,0 +1,6 @@
+module EqualsMicroPerfAndCodeGenerationTests = 
+
+    [<Struct>]
+    type SomeGenericUnion<'T> = SomeGenericUnion of 'T * int
+
+    SomeGenericUnion (1, 2) = SomeGenericUnion (2, 3) |> ignore