--- a/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals15.fsx
+++ b/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals15.fsx
@@ -1,0 +1,6 @@
+module EqualsMicroPerfAndCodeGenerationTests = 
+
+    [<Struct>]
+    type SomeGenericRecord<'T> = { V: 'T; U: int }
+
+    { V = 1; U = 2 } = { V = 2; U = 3 } |> ignore