--- a/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals16.fsx
+++ b/tests/FSharp.Compiler.ComponentTests/EmittedIL/GenericComparison/Equals16.fsx
@@ -1,0 +1,9 @@
+module EqualsMicroPerfAndCodeGenerationTests = 
+
+    [<Struct>]
+    // exact Equals with comparer should be generated internal
+    type internal SomeStruct(v: int, u: int) =
+        member _.V = v
+        member _.U = u
+
+    SomeStruct(1, 2) = SomeStruct(2, 3) |> ignore