@auduchinok 2024-04-16
> This also applies to struct unions in the same fashion:

Should it also cover struct records?