@psfinaki 2024-04-18
> > This also applies to struct unions in the same fashion:
> 
> Should it also cover struct records?

@auduchinok yep, it does. At that point, I just hadn't finished with updating the PR description. Now it's there :)