@auduchinok 2024-05-13
@psfinaki Am I missing something, or this change isn't guarded by a language version check?