@psfinaki 2024-05-14
No, there was a discussion and eventually we decided not to put it here. I am trying to remember the motivation though.