@abelbraaksma 2024-05-16
This PR seems to address this draft RFC, even though that RFC suggested a slightly different approach (that is, it also addresses the `nan <> nan = true` issue).

I'm not quite sure how much overlap there is between that one and this PR, perhaps we should align them so that we can publish this amazing PR with the accompanying RFC.