@psfinaki 2024-05-17
@abelbraaksma thanks for your nice words :)

So yeah this hasn't touched the existing NaN inconsistency - it is still inconsistent. The approach taken in this PR is also substantially different from previous attempts to reduce boxing, but that's because those attempts tried to remove boxing in a different place - basically to use a less generic (and less boxing) comparer. This was finally done earlier this year [here](https://github.com/dotnet/fsharp/pull/16615) - although not including all the original ideas so there is much more to be done in that space.

Anyway, the issue which is focused on the NaN problem is [here](https://github.com/dotnet/fsharp/issues/556), I added some summary there now.